refactor: Move pagination logic to a custom hook
