+++ src/components/shared/MobilePagination/MobilePagination.tsx
@@ -1,9 +1,7 @@
-'use client';
-
-import { useRouter, useSearchParams } from 'next/navigation';
+'use client'
 import { IoIosArrowRoundForward } from 'react-icons/io';
 import Typography from '@components/components/Typography/Typography';
-import { useFilterSearchParams } from '@components/hooks';
+import { usePagination } from '@components/hooks';
 
 import styles from './MobilePagination.module.scss';
 
@@ -12,18 +10,7 @@ interface MobilePaginationProps {
 }
 
 const MobilePagination = ({ totalPages }: MobilePaginationProps) => {
-  const searchParams = useSearchParams();
-  const router = useRouter();
-  const { queryParams } = useFilterSearchParams();
-
-  const page =
-    typeof searchParams.get('page') === 'string'
-      ? Number(searchParams.get('page'))
-      : 1;
-
-  const handlePageClick = (pageNumber: number) => {
-    router.push(`?page=${pageNumber}&perPage=9&${queryParams}#candles-section`);
-  };
+  const { handlePageClick, page } = usePagination();
 
   return (
     <>
