+++ src/components/shared/Pagination/Pagination.tsx
@@ -1,8 +1,7 @@
 'use client';
-import { useRouter, useSearchParams } from 'next/navigation';
 import Container from '@components/components/Container/Container';
 import Section from '@components/components/Section/Section';
-import { useFilterSearchParams } from '@components/hooks';
+import { usePagination } from '@components/hooks';
 
 import styles from './Pagination.module.scss';
 
@@ -11,23 +10,12 @@ interface PaginationProps {
 }
 
 const Pagination = ({ totalPages }: PaginationProps) => {
-  const searchParams = useSearchParams();
-  const router = useRouter();
-  const { queryParams } = useFilterSearchParams();
-
-  const page =
-    typeof searchParams.get('page') === 'string'
-      ? Number(searchParams.get('page'))
-      : 1;
+  const { handlePageClick, page } = usePagination();
 
   const pageRange = Array.from(
     { length: totalPages ?? 1 },
     (_, index) => index + 1
   );
-
-  const handlePageClick = (pageNumber: number) => {
-    router.push(`?page=${pageNumber}&perPage=9&${queryParams}#candles-section`);
-  };
 
   return (
     <Section className={styles.section}>
