+++ src/hooks/index.ts
@@ -1,4 +1,5 @@
 export { useCandleParam } from './useCandleParam';
 export { useFilterSearchParams } from './useFilterSearchParams';
 export { useLangFromPathname } from './useLangFromPathname';
+export { usePagination } from './usePagination';
 export { useProductList } from './useProductList';
